split PsContact out of models
