+++ osidb/tests/factories.py
@@ -11,7 +11,7 @@ from pytz import UTC
 
 from osidb.constants import AFFECTEDNESS_VALID_RESOLUTIONS, DATETIME_FMT
 from osidb.core import generate_acls
-from osidb.dmodels import PsModule, PsProduct, PsUpdateStream
+from osidb.dmodels import PsContact, PsModule, PsProduct, PsUpdateStream
 from osidb.models import (
     CVSS,
     Affect,
@@ -26,7 +26,6 @@ from osidb.models import (
     Impact,
     Package,
     PackageVer,
-    PsContact,
     Snippet,
     SpecialConsiderationPackage,
     Tracker,
